pyproject.toml
[project]
name = "leo"
version = "0.1.0"
description = "GPU Performance Root Cause Analyzer - Python implementation of GPA"
requires-python = ">=3.11"
dependencies = [
    "hpcanalysis",
    "pandas>=2.0.0",
    "click>=8.0.0",
    "pyelftools>=0.29",
    "msgpack>=1.0.0",
    "matplotlib>=3.7.0",
    "openai>=1.0.0",
    "betterbib>=8.0.0b3",
    "python-pptx>=1.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/leo"]

[dependency-groups]
dev = [
    "meson-python>=0.14.0",
    "meson>=1.2.0",
    "ninja>=1.11.0",
    "pybind11>=2.6.0",
    "pytest>=7.0.0",
    "huggingface-hub>=0.20.0",
]

# NVIDIA PyPI for RAPIDS/CUDA packages
[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.uv.sources]
hpcanalysis = { path = "3rdparty/hpcanalysis", editable = false }
